requirements.txt
asgiref==3.7.2
Django==4.2.8
django-filter==23.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.9.0
